src/app/features/collection/guards/check-product-search-params.guard.ts
import { Injectable } from '@angular/core';
import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root',
})
export class CheckProductSearchParamsGuard implements CanActivate {
  canActivate(
    route: ActivatedRouteSnapshot,
    state: RouterStateSnapshot
  ): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
    if (Object.keys(route.queryParams).length > 0) {
      return true;
    } else {
      this._router.navigate(['/home']);
      return false;
    }
  }

  constructor(private _router: Router) {}
}
